Created files for Trainings and Workouts
- Added multiple files for training and workout tables. Files are a
rough outline of what may be needed or created in future versions, to
have a better grasp on what the full project will look like.
- Added a route for workouts in the App
- Added a link in the Header for Workouts to pull all workouts. This may
not be the final home displaying all workouts but is used to help
visualize if it is working as intended.
- Created a index for all workouts which maps over workouts and then
exercises to display relevant data.