Refine look of site and links
- Removed the Nav Link in Header for creating an exercise and placed
a button within the exercise list that links to the creating a exercise
route.
- Removed the temporary 'p' elements when mapping the output in the
'Randomize' component to display line breaks for each value, and a break
between each ul mapping.
- Code has been added to workouts that is planned to link to creating a
workout form. Currently this code is commented out as it is still a work
progress and not fully functioning the way it is intended to.