Save randomized workout feature added
Completed a feature where the list of randomized exercises could be
added to a workout via the trainings (joins) table.

- An input for date added to page
- Save workout button added and will run a function to save the list of
randomized exercises, creating a full workout in the process.

Note: Still need to add a function that will output an error message if
there are no exercises being displayed, or only create the 'Save' form
after the randomize button has been clicked.